Refactor check_for_toots, move code to different class This leaves only daemon code in check_for_toots, move processing onto a separate class to allow testing.
